refactor(iter1-sol): General code cleanup
Specifically, the following tools give clean output (mostly). The remaining issues are false positives.
 - g++ with full warnings enabled (-W -Wall -Wextra -Weffc++)
 - clang-check with full warnings enable (-Weverything)
 - cppcheck with full warnings enabled

However, clang-tidy when set to the google C++ style still reports a LOT of stuff that we need to review.

Footer: Provide links to any relevant tickets, articles, or other resources

+++ src/ColorData.cpp
@@ -51,7 +51,9 @@ float ColorData::getAlpha() const {
 }
 
 float ColorData::getLuminance() const {
-        return 0.2126*m_red + 0.7152*m_green + 0.0722*m_blue;
+    return static_cast<float>(0.2126)*m_red +
+        static_cast<float>(0.7152)*m_green +
+        static_cast<float>(0.0722)*m_blue;
 }
 
 ColorData ColorData::clampedColor() const {
